fix: enum generation issues
