+++ src/download-helpers.ts
@@ -9,7 +9,7 @@ import rimraf from 'rimraf'
 import tar from 'tar'
 
 import { buildObjectFromEntries } from './ts-helpers.js'
-import { GeoIpDbName, Path, Checksum } from './index.js'
+import { GeoIpDbName, Path, Checksum } from './primitives.js'
 
 const REDIST_MIRROR_URL = 'https://raw.githubusercontent.com/GitSquared/node-geolite2-redist/master/redist/'
